services/api/requirements.txt
boto3>=1.34
pydantic>=2.6
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
aws-lambda-powertools>=2.36
